requirements.txt
python-docx==0.8.6
MySQL-python==1.2.5
Pillow==4.0.0
django-suit==0.2.16
gunicorn==19.4.5
requests